+++ django_mobile_app_distribution/forms.py
@@ -9,6 +9,7 @@ from django.utils.translation import ugettext_lazy as _, string_concat
 
 import settings as app_dist_settings
 from models import IosApp
+import posixpath
 
 
 log = logging.getLogger(__name__)
@@ -42,19 +43,26 @@ class IosAppAdminForm(AppAdminForm):
             filename = hex + '.ipa'
             self.fields['file_name'].initial = filename
 
+
+        # self.fields['file_name'].help_text = string_concat(
+        #     '<span style=color:red;>',
+        #     _("Don't change this file name!"),
+        #     '</span><br />',
+        #     _('Paste the file name and Ad Hoc URL into the Xcode Enterprise Ad Hoc dialog.'),
+        #     '<br />',
+        #     _('File name:'),
+        #     ' ',
+        #     '<span style=color:red;>',
+        #     filename,
+        #     '</span><br />',
+        #     _('Ad Hoc URL:'),
+        #     ' ',
+        #     '<span style=color:red;>',
+        #     '/'.join(s.strip('/') for s in (Site.objects.get_current().domain, settings.MEDIA_URL, app_dist_settings.MOBILE_APP_DISTRIBUTION_IOS_UPLOAD_TO_DIRECTORY_NAME, filename)),
+        #     '</span>')
+
         self.fields['file_name'].help_text = string_concat(
             '<span style=color:red;>',
             _("Don't change this file name!"),
-            '</span><br />',
-            _('Paste the file name and Ad Hoc URL into the Xcode Enterprise Ad Hoc dialog.'),
-            '<br />',
-            _('File name:'),
-            ' ',
-            '<span style=color:red;>',
-            filename,
-            '</span><br />',
-            _('Ad Hoc URL:'),
-            ' ',
-            '<span style=color:red;>',
-            '/'.join(s.strip('/') for s in (Site.objects.get_current().domain, settings.MEDIA_URL, app_dist_settings.MOBILE_APP_DISTRIBUTION_IOS_UPLOAD_TO_DIRECTORY_NAME, filename)),
-            '</span>')
+            '</span>'
+        )