+++ django_mobile_app_distribution/models.py
@@ -1,6 +1,8 @@
+# -*- coding: utf-8 -*-
 import logging
+from urlparse import urljoin
 
-import django_mobile_app_distribution.settings as app_dist_settings
+from django.core.urlresolvers import reverse
 from django.contrib.auth.models import User, Group
 from django.contrib.sites.models import Site
 from django.core.files.storage import FileSystemStorage
@@ -8,6 +10,7 @@ from django.db import models
 from django.db.models.signals import post_save
 from django.utils.translation import ugettext_lazy as _
 
+import django_mobile_app_distribution.settings as app_dist_settings
 from exceptions import MobileAppDistributionConfigurationException
 
 
@@ -44,7 +47,8 @@ class IosApp(App):
     file_name = models.CharField(max_length=200, verbose_name=_('File name'))
     operating_system = models.CharField( max_length=50, choices=app_dist_settings.OS_CHOICES, default=app_dist_settings.IOS, verbose_name=_('Operating system'), editable=False)
     app_binary = models.FileField(upload_to=app_dist_settings.MOBILE_APP_DISTRIBUTION_IOS_UPLOAD_TO_DIRECTORY_NAME, verbose_name=_('Ad Hoc ipa file'))
-    app_plist = models.FileField(upload_to=app_dist_settings.MOBILE_APP_DISTRIBUTION_IOS_UPLOAD_TO_DIRECTORY_NAME, verbose_name=_('Ad Hoc plist'))
+    app_plist = models.FileField(upload_to=app_dist_settings.MOBILE_APP_DISTRIBUTION_IOS_UPLOAD_TO_DIRECTORY_NAME, verbose_name=_('Ad Hoc plist'), blank=True)
+    bundle_identifier = models.CharField(max_length=200, verbose_name=_('Bundle identifier'), default='')
 
     def get_binary_url(self):
         if not self.app_binary:
@@ -55,23 +59,30 @@ class IosApp(App):
         except Exception:
             raise MobileAppDistributionConfigurationException("The site framework's domain name is used to generate the plist and binary links.  Please configure your current site properly. Also make sure that the SITE_ID in your settings file matches the primary key of your current site.")
 
-        return Site.objects.get_current().domain + self.app_binary.url
+        return urljoin(Site.objects.get_current().domain, self.app_binary.url)
 
 
     def get_plist_url(self):
-        if not self.app_plist:
-            return None
+        """
+        Starting in version 0.3 the app_plist field is no longer populated but dynamically generated.
+        That's because Xcode 6 no longer provides a plist.
+        """
         Site.objects.clear_cache()
+        current_site = None
         try:
-            Site.objects.get_current()
+            current_site = Site.objects.get_current()
         except Exception:
             raise MobileAppDistributionConfigurationException("The site framework's domain name is used to generate the plist and binary links.  Please configure your current site properly. Also make sure that the SITE_ID in your settings file matches the primary key of your current site.")
-        return Site.objects.get_current().domain + self.app_plist.url
+
+        if self.app_plist:
+            return urljoin(current_site.domain, self.app_plist.url)
+        else:
+            return urljoin(current_site.domain, reverse('django_mobile_app_distribution_ios_app_plist', kwargs={'app_id': self.pk}))
 
     class Meta:
         verbose_name = _('iOS App')
         verbose_name_plural = _('iOS Apps')
-        ordering = ( 'name', 'operating_system', '-version', '-build_date',)
+        ordering = ('name', 'operating_system', '-version', '-build_date',)
 
 
 fs = FileSystemStorage(location=app_dist_settings.MOBILE_APP_DISTRIBUTION_ANDROID_FILE_STORAGE_PATH)
@@ -98,3 +109,4 @@ def create_user_info(sender, instance, created, **kwargs):
             pass
 
 post_save.connect(create_user_info, sender=User)
+